+++ src/modules/react-redux-hash-router4/RouterSelectors.js
@@ -1,7 +1,7 @@
 import RouterConstants from './RouterConstants'
 
-const routerState = (state) => state[RouterConstants.name]
-const getLocation = (state) => routerState(state)['locationBeforeTransitions']
+const routerState = (state) => state[RouterConstants.REDUCER_NAME]
+const getLocation = (state) => routerState(state).location
 const getPathName = (state) => getLocation(state).pathname
 const getSearch = (state) => getLocation(state).search
 const getHash = (state) => getLocation(state).hash