+++ src/modules/react-redux-hash-router4/util/GetSearchObject.js
@@ -1,0 +1,9 @@
+import QueryString from 'query-string'
+
+/**
+ * helper method for getting query object, utilicing the query-string npm module
+ * @returns {*}
+ */
+export default(queryString) => {
+    return QueryString.parse(queryString)
+}