+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [unreleased]
 
 [0.6.0-rc-x]
+- ufpdate ufp-core router4 and, redux4 update, remove getState() method completely from any action payload
+- change getState() to globalState in ufp-middleware actions - redux4.0
 - async router dev update 
-- karma and mocha updated to 2 and 5
+- karma and mocha updated to 2 and 5  	
 - replace lodash with lodash-es
 - update react 16.3.1
 - remove flow dependency